@@ -1,29 +1,33 @@
-const rollup = require('rollup')
-const typescript = require("rollup-plugin-typescript2")
-const expect = require('chai').expect
-const fs = require('fs')
-const path = require('path')
-const git = require('git-rev-sync')
-const clear = require("rollup-plugin-clear");
-const copy = require("rollup-plugin-copy")
+import * as rollup from 'rollup';
+import typescript from "rollup-plugin-typescript2";
+import { expect } from 'chai';
+import { describe, it } from 'mocha';
+import fs from 'fs';
+import path from 'path';
+import git from 'git-rev-sync';
+import del from "rollup-plugin-delete";
+import copy from "rollup-plugin-copy";
+import commonjs from '@rollup/plugin-commonjs';
 
-const screeps = require('../src/rollup-plugin-screeps')
+
+import * as screeps from '../src/rollup-plugin-screeps';
 
 
-describe('Rollup Screeps Plugin', function(){
+describe('Rollup Screeps Plugin', function () {
+  var config: Partial<screeps.ScreepsConfig>;
   it('should support tokens for screeps.com and email/password for any other server', () => {
-    var config = {
-      "token": "foo",
-      "branch": "auto",
-      "protocol": "https",
-      "hostname": "screeps.com",
-      "port": 443,
-      "path": "/"
+    var config: Partial<screeps.ScreepsConfig> = {
+      token: "foo",
+      branch: "auto",
+      protocol: 'https',
+      hostname: "screeps.com",
+      port: 443,
+      path: "/"
     }
 
     expect(screeps.validateConfig(config)).to.equal(true)
 
-    var config = {
+    var config: Partial<screeps.ScreepsConfig> = {
       "email": "you@domain.tld",
       "password": "foo",
       "branch": "auto",
@@ -35,7 +39,7 @@ describe('Rollup Screeps Plugin', function(){
 
     expect(screeps.validateConfig(config)).to.equal(false)
 
-    var config = {
+    var config: Partial<screeps.ScreepsConfig> = {
       "token": "foo",
       "branch": "auto",
       "protocol": "https",
@@ -46,7 +50,7 @@ describe('Rollup Screeps Plugin', function(){
 
     expect(screeps.validateConfig(config)).to.equal(true)
 
-    var config = {
+    var config: Partial<screeps.ScreepsConfig> = {
       "email": "you@domain.tld",
       "password": "foo",
       "branch": "auto",
@@ -59,7 +63,7 @@ describe('Rollup Screeps Plugin', function(){
     expect(screeps.validateConfig(config)).to.equal(true)
   })
 
-  it('should generate source maps', async function(){
+  it('should generate source maps', async function () {
     var options = {
       input: './tests/fixtures/main.ts',
       output: {
@@ -68,24 +72,23 @@ describe('Rollup Screeps Plugin', function(){
         format: 'cjs'
       },
       plugins: [
-        clear({"targets": ["./tests/dist"]}),
-        typescript({tsconfig: './tests/tsconfig.json'}),
-        screeps.screeps({dryRun: true})
+        del({ "targets": ["./tests/dist"] }),
+        commonjs(),
+        typescript({ tsconfig: './tests/tsconfig.json' }),
+        screeps.screeps({ dryRun: true })
       ]
     }
 
-    let bundle = await rollup.rollup(options);
-    let output = await bundle.write(options.output);
-    
-    // Iterate through bundle and test if type===chunk && map is defined
-    let itemName;
-    for (itemName in bundle) {
-      let item = bundle[itemName];
-      if (item.type === "chunk" && item.map) {
+    let bundle = await rollup.rollup(options as rollup.RollupOptions);
+    let output = (await bundle.write(options.output as rollup.OutputOptions)).output;
 
-        expect(item.map.toString()).to.match(/^module.exports/)
+    output.forEach(item => {
+      if (item.type !== undefined && item.type === "chunk" && output.map !== undefined) {
+        expect(output.map.toString(), output.map.toString()).to.match(/^module.exports/)
+        console.log(output.map.toString())
       }
-    }
+    });
+
     var basePath = path.join(__dirname, 'dist')
     var originalPath = path.join(basePath, 'main.js.map')
     var newPath = path.join(basePath, 'main.js.map.js')
@@ -95,7 +98,7 @@ describe('Rollup Screeps Plugin', function(){
 
   })
 
-  it('should generate branch name', async function(){
+  it('should generate branch name', async function () {
     var screepsOptions = {
       dryRun: true
     }
@@ -108,19 +111,20 @@ describe('Rollup Screeps Plugin', function(){
         format: 'cjs'
       },
       plugins: [
-        clear({"targets": ["./tests/dist"]}),
-        typescript({tsconfig: './tests/tsconfig.json'}),
+        del({ "targets": ["./tests/dist"] }),
+        typescript({ tsconfig: './tests/tsconfig.json' }),
+        commonjs(),
         screeps.screeps(screepsOptions)
       ]
     }
 
-    let bundle = await rollup.rollup(options);
-    let output = await bundle.generate(options.output);
+    let bundle = await rollup.rollup(options as rollup.RollupOptions);
+    let output = await bundle.write(options.output as rollup.OutputOptions);
 
     expect(screeps.getBranchName('auto')).to.equal(git.branch())
   })
 
-  it('should use the branch name', async function(){
+  it('should use the branch name', async function () {
     var screepsOptions = {
       dryRun: true
     }
@@ -133,19 +137,20 @@ describe('Rollup Screeps Plugin', function(){
         format: 'cjs'
       },
       plugins: [
-        clear({"targets": ["./tests/dist"]}),
-        typescript({tsconfig: './tests/tsconfig.json'}),
+        del({ "targets": ["./tests/dist"] }),
+        typescript({ tsconfig: './tests/tsconfig.json' }),
+        commonjs(),
         screeps.screeps(screepsOptions)
       ]
     }
 
-    let bundle = await rollup.rollup(options);
-    let output = await bundle.generate(options.output);
+    let bundle = await rollup.rollup(options as rollup.RollupOptions);
+    let output = await bundle.write(options.output as rollup.OutputOptions);
 
     expect(screeps.getBranchName('ai')).to.equal('ai')
   })
 
-  it('should create a list of files to upload', async function(){
+  it('should create a list of files to upload', async function () {
     var screepsOptions = {
       dryRun: true
     }
@@ -158,8 +163,9 @@ describe('Rollup Screeps Plugin', function(){
         format: 'cjs'
       },
       plugins: [
-        clear({"targets": ["./tests/dist"]}),
-        typescript({tsconfig: './tests/tsconfig.json'}),
+        del({ "targets": ["./tests/dist"] }),
+        typescript({ tsconfig: './tests/tsconfig.json' }),
+        commonjs(),
         copy({
           targets: [
             { src: "./tests/fixtures/*.wasm", dest: "./tests/dist" }
@@ -169,18 +175,18 @@ describe('Rollup Screeps Plugin', function(){
       ]
     }
 
-    let bundle = await rollup.rollup(options);
-    let output = await bundle.write(options.output);
-    
+    let bundle = await rollup.rollup(options as rollup.RollupOptions);
+    let output = await bundle.write(options.output as rollup.OutputOptions);
+
     var code = screeps.getFileList(options.output.file)
 
     expect(Object.keys(code).length).to.equal(3)
     expect(code.main).to.match(/input/)
     expect(code['main.js.map']).to.match(/^module.exports/)
- 
+
   })
 
-  it('should upload WASM files as binary modules', async function() {
+  it('should upload WASM files as binary modules', async function () {
     var screepsOptions = {
       dryRun: true
     }
@@ -193,8 +199,9 @@ describe('Rollup Screeps Plugin', function(){
         format: 'cjs'
       },
       plugins: [
-        clear({"targets": ["./tests/dist"]}),
-        typescript({tsconfig: './tests/tsconfig.json'}),
+        del({ "targets": ["./tests/dist"] }),
+        typescript({ tsconfig: './tests/tsconfig.json' }),
+        commonjs(),
         copy({
           targets: [
             { src: "./tests/fixtures/*.wasm", dest: "./tests/dist" }
@@ -204,17 +211,17 @@ describe('Rollup Screeps Plugin', function(){
       ]
     }
 
-    let bundle = await rollup.rollup(options);
-    let output = await bundle.write(options.output);
-    
+    let bundle = await rollup.rollup(options as rollup.RollupOptions);
+    let output = await bundle.write(options.output as rollup.OutputOptions);
+
     var code = screeps.getFileList(options.output.file)
 
     expect(code['wasm_module.wasm']).to.be.an('object');
-    expect(code['wasm_module.wasm'].binary).to.be.a('string')
+    expect((code['wasm_module.wasm'] as screeps.BinaryModule).binary).to.be.a('string')
     expect(code.main).to.be.a('string')
   })
 
-  it('should get the config', function(){
+  it('should get the config', function () {
     var config = screeps.loadConfigFile('./tests/fixtures/screeps.json')
     expect(config.branch).to.equal('foo')
   })
