@@ -1,5 +1,6 @@
-import clear from "rollup-plugin-clear";
+import del from "rollup-plugin-delete";
 import ts from "rollup-plugin-typescript2";
+import commonjs from '@rollup/plugin-commonjs';
 
 const pkg = require("./package.json");
 
@@ -16,7 +17,8 @@ export default {
     }
   ],
 	plugins: [
-    clear({"targets":["dist/"]}),
+    del({"targets":["dist/"]}),
+    commonjs(),
 		ts()
 	],
   external: [ "screeps-api", "git-rev-sync", "fs", "path" ]
